# internal: Store function param names in ItemTree