@Veykril 2021-12-20
integrated benchmark results
Before:
```
  390ms - unqualified path completion
       89ms - CompletionContext::new
            9ms - parse_query @ FileId(178)
           19ms - descend_into_macros
               19ms - Semantics::analyze_impl
                   18ms - SourceBinder::to_module_def
                       18ms - crate_def_map:wait
                            5ms - file_item_tree_query (1 calls)
                           12ms - ???
                        0   - relevant_crates (1 calls)
                    0   - body_with_source_map_query (1 calls)
                    0   - crate_def_map:wait (2 calls)
                    0   - impl_data_query (1 calls)
           58ms - Semantics::analyze_impl
               58ms - infer:wait @ name
                   57ms - infer_query
                       56ms - deref
                           56ms - deref_by_trait
                               56ms - trait_solve::wait
                                    0   - impl_data_query (65 calls)
                                   56ms - ???
                        0   - crate_def_map:wait (32 calls)
                        0   - deref (2 calls)
                        0   - deref_by_trait (1 calls)
                        0   - resolve_obligations_as_possible (21 calls)
                        0   - trait_solve::wait (13 calls)
                    0   - crate_def_map:wait (3 calls)
                0   - SourceBinder::to_module_def (1 calls)
                0   - crate_def_map:wait (1 calls)
            0   - Semantics::analyze_impl (2 calls)
            0   - SourceBinder::to_module_def (2 calls)
            0   - crate_def_map:wait (1 calls)
       61ms - complete_unqualified_path
            0   - crate_def_map:wait (226 calls)
            0   - item::Builder::build (3 calls)
           44ms - parse_query (37 calls)
            6ms - render_resolution (350 calls)
           10ms - ???
       25ms - iterate_method_candidates
           10ms - trait_solve::wait
               10ms - trait_solve_query @ Ord
                    0   - crate_def_map:wait (1 calls)
                   10ms - impl_datum (14 calls)
            0   - crate_def_map:wait (725 calls)
            0   - deref (1 calls)
            0   - generic_params_query (218 calls)
            6ms - render_method (28 calls)
            4ms - trait_solve::wait (73 calls)
      214ms - import_on_the_fly @ sel
            8ms - import_assets::search_for_imports
                8ms - import_assets::search_for
                    8ms - import_assets::path_applicable_imports
                        7ms - items_with_name @ Name: sel, crate: Exclude, assoc items: Some("hir"), limit: Some(40)
                            7ms - find_items
                                5ms - crate_symbols (1 calls)
                                2ms - query_external_importables (1 calls)
                        1ms - find_path_prefixed (46 calls)
                        0   - get_name_definition (40 calls)
                    0   - import_assets::scope_definitions (1 calls)
          135ms - render_resolution
              135ms - render_fn
                  134ms - parse_query @ FileId(5088)
                    0   - crate_def_map:wait (7 calls)
                    0   - generic_params_query (2 calls)
                    0   - item::Builder::build (1 calls)
           47ms - render_resolution
               47ms - render_fn
                   47ms - parse_query @ FileId(5086)
                    0   - crate_def_map:wait (7 calls)
                    0   - generic_params_query (2 calls)
                    0   - item::Builder::build (1 calls)
           11ms - render_resolution
               11ms - render_fn
                   10ms - parse_query @ FileId(5114)
                    0   - crate_def_map:wait (5 calls)
                    0   - generic_params_query (1 calls)
                    0   - item::Builder::build (1 calls)
            0   - crate_def_map:wait (4 calls)
            0   - parse_query (1 calls)
           10ms - render_resolution (36 calls)
        0   - crate_def_map:wait (6 calls)
        0   - deref (1 calls)
        0   - infer:wait (1 calls)
        0   - item::Builder::build (12 calls)
  102ms - dot completion
       90ms - CompletionContext::new
            9ms - parse_query @ FileId(178)
           19ms - descend_into_macros
               19ms - Semantics::analyze_impl
                   19ms - SourceBinder::to_module_def
                       19ms - crate_def_map:wait
                            6ms - file_item_tree_query @ HirFileId(FileId(FileId(178)))
                           13ms - ???
                        0   - relevant_crates (1 calls)
                    0   - body_with_source_map_query (1 calls)
                    0   - crate_def_map:wait (2 calls)
                    0   - impl_data_query (1 calls)
           59ms - Semantics::analyze_impl
               59ms - infer:wait @ name
                   57ms - infer_query
                       56ms - deref_by_trait
                           56ms - trait_solve::wait
                                0   - impl_data_query (65 calls)
                               56ms - ???
                        0   - crate_def_map:wait (31 calls)
                        0   - deref (3 calls)
                        0   - deref_by_trait (1 calls)
                        0   - resolve_obligations_as_possible (23 calls)
                        0   - trait_solve::wait (13 calls)
                    0   - crate_def_map:wait (3 calls)
                0   - SourceBinder::to_module_def (1 calls)
                0   - crate_def_map:wait (1 calls)
            0   - Semantics::analyze_impl (2 calls)
            0   - SourceBinder::to_module_def (2 calls)
            0   - crate_def_map:wait (1 calls)
        8ms - import_on_the_fly @
            8ms - import_assets::search_for_imports
                8ms - import_assets::search_for
                    8ms - import_assets::trait_applicable_items
                        0   - applicable_inherent_traits (1 calls)
                        0   - deref (2 calls)
                        0   - env_traits (1 calls)
                        0   - get_name_definition (40 calls)
                        4ms - items_with_name (1 calls)
                        2ms - iterate_method_candidates (1 calls)
                    0   - import_assets::scope_definitions (1 calls)
            0   - Semantics::analyze_impl (1 calls)
        0   - Semantics::analyze_impl (2 calls)
        0   - complete_unqualified_path (1 calls)
        0   - crate_def_map:wait (70 calls)
        0   - deref (1 calls)
        0   - item::Builder::build (10 calls)
        3ms - iterate_method_candidates (1 calls)
        0   - lang_item_query (1 calls)
        0   - trait_solve::wait (2 calls)
```
After:
```
  176ms - unqualified path completion
       90ms - CompletionContext::new
            9ms - parse_query @ FileId(178)
           19ms - descend_into_macros
               19ms - Semantics::analyze_impl
                   19ms - SourceBinder::to_module_def
                       19ms - crate_def_map:wait
                            6ms - file_item_tree_query @ HirFileId(FileId(FileId(178)))
                           13ms - ???
                        0   - relevant_crates (1 calls)
                    0   - body_with_source_map_query (1 calls)
                    0   - crate_def_map:wait (2 calls)
                    0   - impl_data_query (1 calls)
           59ms - Semantics::analyze_impl
               59ms - infer:wait @ name
                   57ms - infer_query
                       56ms - deref
                           56ms - deref_by_trait
                               56ms - trait_solve::wait
                                    0   - impl_data_query (65 calls)
                                   56ms - ???
                        0   - crate_def_map:wait (32 calls)
                        0   - deref (2 calls)
                        0   - deref_by_trait (1 calls)
                        0   - resolve_obligations_as_possible (21 calls)
                        0   - trait_solve::wait (13 calls)
                    0   - crate_def_map:wait (3 calls)
                0   - SourceBinder::to_module_def (1 calls)
                0   - crate_def_map:wait (1 calls)
            0   - Semantics::analyze_impl (2 calls)
            0   - SourceBinder::to_module_def (2 calls)
            0   - crate_def_map:wait (1 calls)
       55ms - complete_unqualified_path
            0   - crate_def_map:wait (226 calls)
            0   - item::Builder::build (3 calls)
           44ms - parse_query (37 calls)
            1ms - render_resolution (350 calls)
            9ms - ???
       20ms - iterate_method_candidates
           10ms - trait_solve::wait
               10ms - trait_solve_query @ Ord
                    0   - crate_def_map:wait (1 calls)
                   10ms - impl_datum (14 calls)
            0   - crate_def_map:wait (725 calls)
            0   - deref (1 calls)
            0   - generic_params_query (218 calls)
            0   - render_method (28 calls)
            4ms - trait_solve::wait (73 calls)
       11ms - import_on_the_fly @ sel
            9ms - import_assets::search_for_imports
                9ms - import_assets::search_for
                    9ms - import_assets::path_applicable_imports
                        8ms - items_with_name @ Name: sel, crate: Exclude, assoc items: Some("hir"), limit: Some(40)
                            8ms - find_items
                                6ms - crate_symbols @ Query { query: "sel", lowercased: "sel", only_types: false, libs: false, exact: false, case_sensitive: false, limit: 40 }
                                    0   - crate_def_map:wait (1 calls)
                                    0   - symbol_index::Query::search (1 calls)
                                    6ms - ???
                                2ms - query_external_importables (1 calls)
                        1ms - find_path_prefixed (46 calls)
                        0   - get_name_definition (40 calls)
                    0   - import_assets::scope_definitions (1 calls)
            0   - crate_def_map:wait (4 calls)
            0   - parse_query (1 calls)
            0   - render_resolution (39 calls)
        0   - crate_def_map:wait (6 calls)
        0   - deref (1 calls)
        0   - infer:wait (1 calls)
        0   - item::Builder::build (12 calls)
  101ms - dot completion
       89ms - CompletionContext::new
            9ms - parse_query @ FileId(178)
           19ms - descend_into_macros
               19ms - Semantics::analyze_impl
                   19ms - SourceBinder::to_module_def
                       19ms - crate_def_map:wait
                            6ms - file_item_tree_query @ HirFileId(FileId(FileId(178)))
                           13ms - ???
                        0   - relevant_crates (1 calls)
                    0   - body_with_source_map_query (1 calls)
                    0   - crate_def_map:wait (2 calls)
                    0   - impl_data_query (1 calls)
           59ms - Semantics::analyze_impl
               59ms - infer:wait @ name
                   57ms - infer_query
                       56ms - deref_by_trait
                           56ms - trait_solve::wait
                                0   - impl_data_query (65 calls)
                               56ms - ???
                        0   - crate_def_map:wait (31 calls)
                        0   - deref (3 calls)
                        0   - deref_by_trait (1 calls)
                        0   - resolve_obligations_as_possible (23 calls)
                        0   - trait_solve::wait (13 calls)
                    0   - crate_def_map:wait (3 calls)
                0   - SourceBinder::to_module_def (1 calls)
                0   - crate_def_map:wait (1 calls)
            0   - Semantics::analyze_impl (2 calls)
            0   - SourceBinder::to_module_def (2 calls)
            0   - crate_def_map:wait (1 calls)
        8ms - import_on_the_fly @
            7ms - import_assets::search_for_imports
                7ms - import_assets::search_for
                    7ms - import_assets::trait_applicable_items
                        0   - applicable_inherent_traits (1 calls)
                        0   - deref (2 calls)
                        0   - env_traits (1 calls)
                        0   - get_name_definition (40 calls)
                        4ms - items_with_name (1 calls)
                        2ms - iterate_method_candidates (1 calls)
                    0   - import_assets::scope_definitions (1 calls)
            0   - Semantics::analyze_impl (1 calls)
        0   - Semantics::analyze_impl (2 calls)
        0   - complete_unqualified_path (1 calls)
        0   - crate_def_map:wait (70 calls)
        0   - deref (1 calls)
        0   - item::Builder::build (10 calls)
        3ms - iterate_method_candidates (1 calls)
        0   - lang_item_query (1 calls)
        0   - trait_solve::wait (2 calls)
```